pyproject.toml
[project]
name = "gpt-command-line"
version = "0.1.4"
description = "Command-line interface for ChatGPT, Claude and Bard"
authors = [{name = "Val Kharitonov", email = "val@kharvd.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
keywords = ["cli", "command-line", "assistant", "openai", "claude", "bard", "gpt-3", "gpt-4", "llm", "chatgpt", "gpt-cli", "google-bard", "anthropic", "gpt-client", "anthropic-claude", "palm2"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic==0.7.7",
    "attrs==23.1.0",
    "black==23.1.0",
    "google-generativeai==0.1.0",
    "openai==1.3.8",
    "prompt-toolkit==3.0.41",
    "pytest==7.3.1",
    "PyYAML==6.0",
    "rich==13.7.0",
    "tiktoken==0.5.2",
    "tokenizers==0.15.0",
    "typing_extensions==4.5.0",
    "pytz==2023.3.post1",
    "jinja2==3.1.2",
]

[project.optional-dependencies]
llama = [
    "llama-cpp-python==0.1.57",
]

[project.urls]
"Homepage" = "https://github.com/kharvd/gpt-cli"

[project.scripts]
gpt = "gptcli.gpt:main"

[build-system]
requires = ["pip>=23.0.0", "setuptools>=58.0.0", "wheel"]
build-backend = "setuptools.build_meta"
